# public/docs/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Autours API",
    "description": "Complete API documentation for Autours car rental platform. Covers all JSON endpoints needed to build a separated frontend.\n\n**Auth modes:**\n- Session cookie auth for web routes (login via `/login` sets session cookie)\n- Bearer token (Sanctum) for API routes (login via `/api/external/supplier/login`)\n- Some routes use middleware guards: `admin`, `admin_or_supplier`, `active_supplier`, `customer`, `member`",
    "version": "2.0.0",
    "contact": {
      "name": "Autours Support",
      "email": "info@autours.net"
    }
  },
  "servers": [
    {
      "url": "https://www.autours.net",
      "description": "Production Server"
    },
    {
      "url": "http://localhost:8000",
      "description": "Local Development Server"
    }
  ],
...
  "paths": {
...
    "/booking/{id}": {
      "get": {
        "tags": [
          "Bookings & Rentals"
        ],
        "summary": "Get booking details",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Booking details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RentalResponse"
                }
              }
            }
          }
        }
      }
    },
    "/invoice/booking/{id}": {
      "get": {
        "tags": [
          "Customer"
        ],
        "summary": "Download booking invoice PDF",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "PDF download",
            "content": {
              "application/pdf": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "Token",
        "description": "Sanctum token from supplier/external login"
      }
    },
    "schemas": {
...
      "Rental": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "order_number": {
            "type": "string",
            "example": "USATR0001"
          },
          "customer_id": {
            "type": "integer",
            "example": 10
          },
          "supplier_id": {
            "type": "integer",
            "example": 5
          },
          "vehicle_id": {
            "type": "integer",
            "example": 1
          },
          "price": {
            "type": "number",
            "example": 350.0
          },
          "supplier_price": {
            "type": "number",
            "example": 300.0
          },
          "profit_margin": {
            "type": "number",
            "example": 50.0
          },
          "start_date": {
            "type": "string",
            "format": "date",
            "example": "2025-06-01"
          },
          "end_date": {
            "type": "string",
            "format": "date",
            "example": "2025-06-07"
          },
          "start_time": {
            "type": "string",
            "example": "10:00:00"
          },
          "end_time": {
            "type": "string",
            "example": "10:00:00"
          },
          "currency": {
            "type": "string",
            "example": "USD"
          },
          "number_of_days": {
            "type": "integer",
            "example": 6
          },
          "order_status": {
            "type": "integer",
            "example": 1
          },
          "rate": {
            "type": "number",
            "example": 4.5
          },
          "old_rental_id": {
            "type": "integer",
            "example": null
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "RentalResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "boolean"
          },
          "data": {
            "$ref": "#/components/schemas/Rental"
          }
        }
      },
...
    }
  }
}